+++ app/questions.js
@@ -1,0 +1,57 @@
+export const questions = [
+    {
+      id: 1,
+      text: "Your phone would expose you for...",
+      options: [
+        "ordering food while ordering food",
+        "setting 8 alarms 3 minutes apart",
+        "googling things mid-conversation",
+        "typing lol but not laughing",
+        "saving posts you'll never look at"
+      ]
+    },
+    {
+      id: 2,
+      text: "Your toxic trait is...",
+      options: [
+        "thinking sales means saving money",
+        "calling my mom counts as therapy",
+        "being early is my personality",
+        "believing in my past life",
+        "dating my ick"
+      ]
+    },
+    {
+      id: 3,
+      text: "You're living in a movie, it's...",
+      options: [
+        "groundhog day but make it naps",
+        "fast and furious but on electric scooters",
+        "inception but it's just daydreaming",
+        "mission impossible: inbox zero",
+        "jurassic park but with cats"
+      ]
+    },
+    {
+      id: 4,
+      text: "Your weekend plans leaked...",
+      options: [
+        "true crime podcast and face mask",
+        "reorganizing my entire life",
+        "becoming a morning person (again)",
+        "hot girl walk to get iced coffee",
+        "booking flights i can't afford"
+      ]
+    },
+    {
+      id: 5,
+      text: "Your screen time report is mostly...",
+      options: [
+        "reddit rabbit holes at 3am",
+        "watching other people cook",
+        "rating dogs in my area",
+        "google maps street view tourism",
+        "scrolling houses i can't afford"
+      ]
+    }
+  ];
